refactor MxmlTupletState
