+++ src/importexport/musicxml/internal/musicxml/importmxmlpass1.cpp
@@ -3102,252 +3102,6 @@ void MusicXMLParserPass1::notations(MxmlStartStop& tupletStartStop)
 }
 
 //---------------------------------------------------------
-//   smallestTypeAndCount
-//---------------------------------------------------------
-
-/**
- Determine the smallest note type and the number of those
- present in a ChordRest.
- For a note without dots the type equals the note type
- and count is one.
- For a single dotted note the type equals half the note type
- and count is three.
- A double dotted note is similar.
- Note: code assumes when duration().type() is incremented,
- the note length is divided by two, checked by tupletAssert().
- */
-
-static void smallestTypeAndCount(const TDuration durType, int& type, int& count)
-{
-    type = int(durType.type());
-    count = 1;
-    switch (durType.dots()) {
-    case 0:
-        // nothing to do
-        break;
-    case 1:
-        type += 1;                 // next-smaller type
-        count = 3;
-        break;
-    case 2:
-        type += 2;                 // next-next-smaller type
-        count = 7;
-        break;
-    default:
-        LOGD("smallestTypeAndCount() does not support more than 2 dots");
-    }
-}
-
-//---------------------------------------------------------
-//   matchTypeAndCount
-//---------------------------------------------------------
-
-/**
- Given two note types and counts, if the types are not equal,
- make them equal by successively doubling the count of the
- largest type.
- */
-
-static void matchTypeAndCount(int& type1, int& count1, int& type2, int& count2)
-{
-    while (type1 < type2) {
-        type1++;
-        count1 *= 2;
-    }
-    while (type2 < type1) {
-        type2++;
-        count2 *= 2;
-    }
-}
-
-//---------------------------------------------------------
-//   addDurationToTuplet
-//---------------------------------------------------------
-
-/**
- Add duration to tuplet duration
- Determine type and number of smallest notes in the tuplet
- */
-
-void MxmlTupletState::addDurationToTuplet(const Fraction dur, const Fraction timeMod)
-{
-    /*
-    LOGD("1 duration %s timeMod %s -> state.tupletType %d state.tupletCount %d state.actualNotes %d state.normalNotes %d",
-           muPrintable(duration.print()),
-           muPrintable(timeMod.print()),
-           m_tupletType,
-           m_tupletCount,
-           m_actualNotes,
-           m_normalNotes
-           );
-    */
-    if (duration <= Fraction(0, 1)) {
-        // first note: init variables
-        actualNotes = timeMod.denominator();
-        normalNotes = timeMod.numerator();
-        smallestTypeAndCount(dur / timeMod, tupletType, tupletCount);
-    } else {
-        int noteType = 0;
-        int noteCount = 0;
-        smallestTypeAndCount(dur / timeMod, noteType, noteCount);
-        // match the types
-        matchTypeAndCount(tupletType, tupletCount, noteType, noteCount);
-        tupletCount += noteCount;
-    }
-    duration += dur;
-    /*
-    LOGD("2 duration %s -> state.tupletType %d state.tupletCount %d state.actualNotes %d state.normalNotes %d",
-           muPrintable(duration.print()),
-           m_tupletType,
-           m_tupletCount,
-           m_actualNotes,
-           m_normalNotes
-           );
-    */
-}
-
-//---------------------------------------------------------
-//   determineTupletFractionAndFullDuration
-//---------------------------------------------------------
-
-/**
- Split duration into two factors where fullDuration is note sized
- (i.e. the denominator is a power of 2), 1/2 < fraction <= 1/1
- and fraction * fullDuration equals duration.
- */
-
-void determineTupletFractionAndFullDuration(const Fraction duration, Fraction& fraction, Fraction& fullDuration)
-{
-    fraction = duration;
-    fullDuration = Fraction(1, 1);
-    // move denominator's powers of 2 from fraction to fullDuration
-    while (fraction.denominator() % 2 == 0) {
-        fraction *= 2;
-        fraction.reduce();
-        fullDuration *= Fraction(1, 2);
-    }
-    // move numerator's powers of 2 from fraction to fullDuration
-    while (fraction.numerator() % 2 == 0) {
-        fraction *= Fraction(1, 2);
-        fraction.reduce();
-        fullDuration *= 2;
-        fullDuration.reduce();
-    }
-    // make sure 1/2 < fraction <= 1/1
-    while (fraction <= Fraction(1, 2)) {
-        fullDuration *= Fraction(1, 2);
-        fraction *= 2;
-    }
-    fullDuration.reduce();
-    fraction.reduce();
-
-    /*
-    Examples (note result when denominator is not a power of two):
-    3:2 tuplet of 1/4 results in fraction 1/1 and fullDuration 1/2
-    2:3 tuplet of 1/4 results in fraction 3/1 and fullDuration 1/4
-    4:3 tuplet of 1/4 results in fraction 3/1 and fullDuration 1/4
-    3:4 tuplet of 1/4 results in fraction 1/1 and fullDuration 1/1
-
-     Bring back fraction in 1/2 .. 1/1 range.
-     */
-
-    if (fraction > Fraction(1, 1) && fraction.denominator() == 1) {
-        fullDuration *= fraction;
-        fullDuration.reduce();
-        fraction = Fraction(1, 1);
-    }
-
-    /*
-    LOGD("duration %s fraction %s fullDuration %s",
-           muPrintable(duration.toString()),
-           muPrintable(fraction.toString()),
-           muPrintable(fullDuration.toString())
-           );
-    */
-}
-
-//---------------------------------------------------------
-//   isTupletFilled
-//---------------------------------------------------------
-
-/**
- Determine if the tuplet is completely filled,
- because either (1) it is at least the same duration
- as the specified number of the specified normal type notes
- or (2) the duration adds up to a normal note duration.
-
- Example (1): a 3:2 tuplet with a 1/4 and a 1/8 note
- is filled if normal type is 1/8,
- it is not filled if normal type is 1/4.
-
- Example (2): a 3:2 tuplet with a 1/4 and a 1/8 note is filled.
- */
-
-static bool isTupletFilled(const MxmlTupletState& state, const TDuration normalType, const Fraction timeMod)
-{
-    UNUSED(timeMod);
-    bool res = false;
-    const int actualNotes = state.actualNotes;
-    /*
-    const auto normalNotes = state.m_normalNotes;
-    LOGD("duration %s normalType %s timeMod %s normalNotes %d actualNotes %d",
-           muPrintable(state.m_duration.toString()),
-           muPrintable(normalType.fraction().toString()),
-           muPrintable(timeMod.toString()),
-           normalNotes,
-           actualNotes
-           );
-    */
-
-    int tupletType = state.tupletType;
-    int tupletCount = state.tupletCount;
-
-    if (normalType.isValid()) {
-        int matchedNormalType  = int(normalType.type());
-        int matchedNormalCount = actualNotes;
-        // match the types
-        matchTypeAndCount(tupletType, tupletCount, matchedNormalType, matchedNormalCount);
-        // ... result scenario (1)
-        res = tupletCount >= matchedNormalCount;
-        /*
-        LOGD("normalType valid tupletType %d tupletCount %d matchedNormalType %d matchedNormalCount %d res %d",
-               tupletType,
-               tupletCount,
-               matchedNormalType,
-               matchedNormalCount,
-               res
-               );
-         */
-    } else {
-        // ... result scenario (2)
-        res = tupletCount >= actualNotes;
-        /*
-        LOGD("normalType not valid tupletCount %d actualNotes %d res %d",
-               tupletCount,
-               actualNotes,
-               res
-               );
-         */
-    }
-    return res;
-}
-
-//---------------------------------------------------------
-//   missingTupletDuration
-//---------------------------------------------------------
-
-Fraction missingTupletDuration(const Fraction duration)
-{
-    Fraction tupletFraction;
-    Fraction tupletFullDuration;
-
-    determineTupletFractionAndFullDuration(duration, tupletFraction, tupletFullDuration);
-    Fraction missing = (Fraction(1, 1) - tupletFraction) * tupletFullDuration;
-
-    return missing;
-}
-
-//---------------------------------------------------------
 //   voiceToInt
 //---------------------------------------------------------
 static int voiceStrIndex(const String& v)
@@ -3368,103 +3122,6 @@ int MusicXMLParserPass1::voiceToInt(const String& voice)
         voiceInt = voiceStrIndex(voice); // Handle the rare but techincally in-spec case of a non-int voice
     }
     return voiceInt;
-}
-
-//---------------------------------------------------------
-//   determineTupletAction
-//---------------------------------------------------------
-
-/**
- Update tuplet state using parse result tupletDesc.
- Tuplets with <actual-notes> and <normal-notes> but without <tuplet>
- are handled correctly.
- TODO Nested tuplets are not (yet) supported.
- */
-
-MxmlTupletFlags MxmlTupletState::determineTupletAction(const Fraction noteDuration,
-                                                       const Fraction timeMod,
-                                                       const MxmlStartStop tupletStartStop,
-                                                       const TDuration normalType,
-                                                       Fraction& missingPreviousDuration,
-                                                       Fraction& missingCurrentDuration)
-{
-    const int actNotes = timeMod.denominator();
-    const int norNotes = timeMod.numerator();
-    MxmlTupletFlags res = MxmlTupletFlag::NONE;
-
-    // check for unexpected termination of previous tuplet
-    if (inTuplet && timeMod == Fraction(1, 1)) {
-        // recover by simply stopping the current tuplet first
-        if (!isTupletFilled(*this, normalType, timeMod)) {
-            missingPreviousDuration = missingTupletDuration(duration);
-            //LOGD("tuplet incomplete, missing %s", muPrintable(missingPreviousDuration.print()));
-        }
-        *this = {};
-        res |= MxmlTupletFlag::STOP_PREVIOUS;
-    }
-
-    // check for obvious errors
-    if (inTuplet && tupletStartStop == MxmlStartStop::START) {
-        LOGD("tuplet already started");
-        // recover by simply stopping the current tuplet first
-        if (!isTupletFilled(*this, normalType, timeMod)) {
-            missingPreviousDuration = missingTupletDuration(duration);
-            //LOGD("tuplet incomplete, missing %s", muPrintable(missingPreviousDuration.print()));
-        }
-        *this = {};
-        res |= MxmlTupletFlag::STOP_PREVIOUS;
-    }
-    if (tupletStartStop == MxmlStartStop::STOP && !inTuplet) {
-        LOGD("tuplet stop but no tuplet started");           // TODO
-        // recovery handled later (automatically, no special case needed)
-    }
-
-    // Tuplet are either started by the tuplet start
-    // or when the time modification is first found.
-    if (!inTuplet) {
-        if (tupletStartStop == MxmlStartStop::START
-            || (!inTuplet && (actNotes != 1 || norNotes != 1))) {
-            if (tupletStartStop != MxmlStartStop::START) {
-                implicit = true;
-            } else {
-                implicit = false;
-            }
-            // create a new tuplet
-            inTuplet = true;
-            res |= MxmlTupletFlag::START_NEW;
-        }
-    }
-
-    // Add chord to the current tuplet.
-    // Must also check for actual/normal notes to prevent
-    // adding one chord too much if tuplet stop is missing.
-    if (inTuplet && !(actNotes == 1 && norNotes == 1)) {
-        addDurationToTuplet(noteDuration, timeMod);
-        res |= MxmlTupletFlag::ADD_CHORD;
-    }
-
-    // Tuplets are stopped by the tuplet stop
-    // or when the tuplet is filled completely
-    // (either with knowledge of the normal type
-    // or as a last resort calculated based on
-    // actual and normal notes plus total duration)
-    // or when the time-modification is not found.
-
-    if (inTuplet) {
-        if (tupletStartStop == MxmlStartStop::STOP
-            || (implicit && isTupletFilled(*this, normalType, timeMod))
-            || (actNotes == 1 && norNotes == 1)) {           // incorrect ??? check scenario incomplete tuplet w/o start
-            if (actNotes > norNotes && !isTupletFilled(*this, normalType, timeMod)) {
-                missingCurrentDuration = missingTupletDuration(duration);
-                LOGD("current tuplet incomplete, missing %s", muPrintable(missingCurrentDuration.toString()));
-            }
-
-            *this = {};
-            res |= MxmlTupletFlag::STOP_CURRENT;
-        }
-    }
-
-    return res;
 }
 
 //---------------------------------------------------------
